requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.0.3
frozenlist==1.7.0
google-ai-generativelanguage==0.6.5
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.7.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.62.3
gspread==6.0.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.0
oauthlib==3.3.0
packaging==25.0
pandas==2.2.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
StrEnum==0.4.15
tqdm==4.67.1
twilio==8.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
